mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "vina-doctor MCP server — exposes medical AI capabilities as MCP tools"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [".."]
include = ["mcp_server*"]

[tool.setuptools.package-dir]
"" = ".."

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "PLC", "PLE", "PLW"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
